Separate reporter utils from bootstrapper
